+++ src/gpu/fastlanes-global.h
@@ -6,12 +6,12 @@
 namespace gpu {
 template <typename T_in, typename T_out>
 void bitunpack_with_function(T_in *__restrict in, T_out *__restrict out,
-                             uint16_t value_bit_width) {
+                             int32_t value_bit_width) {
 	return;
 }
 template <typename T_in, typename T_out>
 void bitunpack_with_reader(T_in *__restrict in, T_out *__restrict out,
-                           uint16_t value_bit_width) {
+                           int32_t value_bit_width) {
 	return;
 }
 } // namespace gpu
